Ora gli indirizzi sono associati agli ordini e non agli user

+++ ApiService/ApiService/Service/AddressBook/AddressBookService.cs
@@ -1,43 +1,32 @@
 ﻿using ApiService.Service.AddressBook.Exceptionz;
 using ApiService.Service.AddressBook.Httpz;
-using ApiService.Service.User.Dto;
-using ApiService.Service.User.Httpz;
+using ApiService.Service.Order.Httpz;
+using ApiService.Service.Product.Exceptionz;
 
 namespace ApiService.Service.AddressBook
 {
     public class AddressBookService : IAddressBookService
     {
+        private readonly IOrderHttpClient _orderHttpClient;
         private readonly IAddressBookHttpClient _addressBookHttpClient;
-        private readonly IUserHttpClient _userHttpClient;
 
         public AddressBookService
         (
-            IAddressBookHttpClient addressBookHttpClient,
-            IUserHttpClient userHttpClient
+            IOrderHttpClient orderHttpClient,
+            IAddressBookHttpClient addressBookHttpClient
         )
         {
-            _userHttpClient = userHttpClient;
+            _orderHttpClient = orderHttpClient;
             _addressBookHttpClient = addressBookHttpClient;
         }
 
         public async Task<bool> Delete(int id)
         {
-            // Se elimino un indirizzo devo eliminare quelli associati agli utenti
-            try
+            // se trovo qualcosa di associato non posso eliminare l'indirizzo
+            var ordersWithAddress = await _orderHttpClient.GetByAddressIdAsync(id);
+            if (ordersWithAddress.Any())
             {
-                var users = await _userHttpClient.GetByAddressId<List<UserDto>>(id);
-                foreach (var user in users)
-                {
-                    var disassociateAddres = (bool)(await _userHttpClient.UpdateAddressId(user.Id, -1));
-                    if (!disassociateAddres)
-                    {
-                        throw new AddressBookException($"Non sono riuscito a disassociare l'address id all'utente {disassociateAddres}");
-                    }
-                }
-            }
-            catch (Exception e)
-            {
-                throw new AddressBookException($"Errore disassociazione address all'user {e.Message}");
+                throw new AddressBookException($"Ordini con id {string.Join(", ", ordersWithAddress.Select(x => x.Id))} associati all'indirizzo, non posso eliminarlo");
             }
 
             return await _addressBookHttpClient.Delete($"{id}");
